Fix: Update `--new-rule` process
Refactor all wizards into a folder.
Refactor shared templates and files.
Fix #857
