--- packages/sonarwhal/src/lib/cli/parsers/templates/config.hbs
@@ -1,13 +1,0 @@
-{
-    "connector": {
-        "name": "jsdom",
-        "options": {
-            "waitFor": 1000
-        }
-    },
-    "formatters": "summary",
-    "rulesTimeout": 120000,
-    "parsers": ["{{normalizedName}}"],
-    "rules": {
-    }
-}
