--- packages/sonarwhal/src/lib/cli/parsers/templates/package.hbs
@@ -1,89 +1,0 @@
-{
-  "author": "",
-  "ava": {
-    "failFast": false,
-    "files": [
-      "dist/tests/**/*.js"
-    ],
-    "timeout": "1m"
-  },
-  "description": "{{description}}",
-  "devDependencies": {
-    "@types/debug": "{{dependencyVersion "@types/debug" "0.0.30"}}",
-    "@types/node": "{{dependencyVersion "@types/node" "8.0.14"}}",
-    "ava": "{{dependencyVersion "ava" "^0.24.0"}}",
-    "cpx": "{{dependencyVersion "cpx" "^1.5.0"}}",
-    "eslint": "{{dependencyVersion "eslint" "^4.12.1"}}",
-    "eslint-plugin-markdown": "{{dependencyVersion "eslint-plugin-markdown" "^1.0.0-beta.7"}}",
-    "eslint-plugin-typescript": "{{dependencyVersion "eslint-plugin-typescript" "^0.8.1"}}",
-    "eventemitter2": "{{dependencyVersion "eventemitter2" "^5.0.0"}}",
-    "markdownlint-cli": "{{dependencyVersion "markdownlint-cli" "^0.5.0"}}",
-    "npm-run-all": "{{dependencyVersion "npm-run-all" "^4.1.2"}}",
-    "nyc": "{{dependencyVersion "nyc" "^11.5.0"}}", 
-    "proxyquire": "{{dependencyVersion "proxyquire" "2.0.0"}}",
-    "rimraf": "{{dependencyVersion "rimraf" "^2.6.2"}}",
-    "sinon": "{{dependencyVersion "sinon" "^4.1.3"}}",
-    "sonarwhal": "^{{version}}",
-    "typescript": "{{dependencyVersion "typescript" "^2.6.2"}}",
-    "typescript-eslint-parser": "{{dependencyVersion "typescript-eslint-parser" "^9.0.1"}}"
-  },
-  "engines": {
-    "node": ">=8.0.0"
-  },
-  "files": [
-    "dist/src",
-    "npm-shrinkwrap.json"
-  ],{{#if official}}
-  "homepage": "https://sonarwhal.com/",{{else}}
-  "homepage": "",{{/if}}
-  "keywords": [
-    "rule",
-    "sonarwhal",
-    "{{normalizedName}}",
-    "{{normalizedName}}-parser"
-  ],
-  "license": "Apache-2.0",
-  "main": "{{packageMain}}",
-  "name": "{{packageName}}",{{#if official}}
-  "nyc": {
-    "extends": "../../../../.nycrc"
-  },{{else}}
-  "nyc": {
-    "branches": 85,
-    "cache": true,
-    "check-coverage": true,
-    "functions": 80,
-    "instrument": true,
-    "lines": 85,
-    "reporter": [
-        "lcov",
-        "text"
-    ],
-    "sourceMap": true,
-    "statements": 85
-  },{{/if}}
-  "scripts": {
-    "build": "npm run clean && npm-run-all build:*",
-    "build:assets": "cpx \"./{src,tests}/**/{!(*.ts),.!(ts)}\" dist",
-    "build:ts": "tsc",  
-    "clean": "rimraf dist",
-    "lint": "npm-run-all lint:*",
-    {{#if official}}
-    "lint:js": "eslint . --cache --ext js --ext md --ext ts --ignore-path ../../.eslintignore --report-unused-disable-directives",{{else}}
-    "lint:js": "eslint . --cache --ext js --ext md --ext ts --report-unused-disable-directives",{{/if}}
-    "lint:md": "markdownlint *.md",
-    "sonarwhal": "node node_modules/sonarwhal/dist/src/bin/sonarwhal.js",
-    "test": "npm run lint && npm run build && nyc ava",
-    "init": "npm install && npm run build",
-    "watch": "npm run build && npm-run-all --parallel -c watch:*",
-    "watch:assets": "npm run build:assets -- -w --no-initial",
-    "watch:test": "ava --watch",
-    "watch:ts": "npm run build:ts -- --watch"
-  },
-  "peerDependencies": {
-    "sonarwhal": "^{{version}}"
-  },{{#if official}}
-  "repository": "sonarwhal/sonarwhal",{{else}}
-  "repository": "",{{/if}}
-  "version": "0.1.0"
-}
