--- packages/sonarwhal/src/lib/cli/rules/templates/new-rule/tsconfig.json.hbs
@@ -1,40 +1,0 @@
-{{#if official}}{
-    "compilerOptions": {
-        "outDir": "dist"
-    },
-    "exclude": [
-        "dist",
-        "node_modules"
-    ],
-    "extends": "../../tsconfig.json",
-    "include": [
-        "src/**/*.ts",
-        "tests/**/*.ts"
-    ]
-}{{else}}{
-    "compilerOptions": {
-        "alwaysStrict": true,
-        "declaration": true,
-        "inlineSourceMap": true,
-        "lib": [
-            "dom",
-            "dom.iterable",
-            "es2017",
-            "esnext",
-            "esnext.asynciterable"
-        ],
-        "module": "commonjs",
-        "newLine": "lf",
-        "outDir": "dist",
-        "removeComments": true,
-        "target": "esnext"
-    },
-    "exclude": [
-        "dist",
-        "node_modules"
-    ],
-    "include": [
-        "src/**/*.ts",
-        "tests/**/*.ts"
-    ]
-}{{/if}}
