+++ packages/sonarwhal/src/lib/cli/wizards/files/.npmrc
@@ -1,0 +1,1 @@
+package-lock=false
