+++ packages/sonarwhal/src/lib/cli/wizards/files/LICENSE.txt
