+++ packages/sonarwhal/src/lib/cli/wizards/no-official-files/.editorconfig
