+++ packages/sonarwhal/src/lib/cli/wizards/shared-templates/package.hbs
@@ -15,7 +15,10 @@
     "cpx": "{{dependencyVersion "cpx" "^1.5.0"}}",
     "eslint": "{{dependencyVersion "eslint" "^4.17.0"}}",
     "eslint-plugin-markdown": "{{dependencyVersion "eslint-plugin-markdown" "^1.0.0-beta.7"}}",
-    "eslint-plugin-typescript": "{{dependencyVersion "eslint-plugin-typescript" "^0.8.1"}}",
+    "eslint-plugin-typescript": "{{dependencyVersion "eslint-plugin-typescript" "^0.8.1"}}",{{#if isParser}}
+    "eventemitter2": "{{dependencyVersion "eventemitter2" "^5.0.0"}}",
+    "proxyquire": "{{dependencyVersion "proxyquire" "2.0.0"}}",
+    "sinon": "{{dependencyVersion "sinon" "^4.1.3"}}",{{/if}}
     "markdownlint-cli": "{{dependencyVersion "markdownlint-cli" "^0.7.0"}}",
     "npm-link-check": "{{dependencyVersion "npm-link-check" "^2.0.0"}}",
     "npm-run-all": "{{dependencyVersion "npm-run-all" "^4.1.2"}}",
@@ -23,7 +26,7 @@
     "rimraf": "{{dependencyVersion "rimraf" "^2.6.2"}}",
     "sonarwhal": "^{{version}}",
     "typescript": "{{dependencyVersion "typescript" "^2.6.2"}}",
-    "typescript-eslint-parser": "{{dependencyVersion "typescript-eslint-parser" "^13.0.0"}}"
+    "typescript-eslint-parser": "{{dependencyVersion "typescript-eslint-parser" "^14.0.0"}}"
   },
   "engines": {
     "node": ">=8.0.0"
@@ -37,8 +40,9 @@
   "keywords": [
     "rule",
     "sonarwhal",
-    "{{normalizedName}}",
-    "{{normalizedName}}-rule"
+    "{{normalizedName}}",{{#if isParser}}
+    "{{normalizedName}}-parser"{{else}}
+    "{{normalizedName}}-rule"{{/if}}
   ],
   "license": "Apache-2.0",
   "main": "{{packageMain}}",
@@ -68,7 +72,7 @@
     "lint": "npm-run-all lint:*",{{#if official}}
     "lint:js": "eslint . --cache --ext js --ext md --ext ts --ignore-path ../../.eslintignore --report-unused-disable-directives",{{else}}
     "lint:js": "eslint . --cache --ext js --ext md --ext ts --report-unused-disable-directives",{{/if}}
-    "lint:md": "markdownlint *.md",
+    "lint:md": "markdownlint --ignore CHANGELOG.md *.md",
     "sonarwhal": "node node_modules/sonarwhal/dist/src/bin/sonarwhal.js",
     "test": "npm run lint && npm run build && nyc ava",
     "init": "npm install && npm run build",
