+++ packages/sonarwhal/src/lib/cli/wizards/shared-templates/tsconfig.json.hbs
