+++ packages/sonarwhal/src/lib/cli/wizards/templates/new-parser/index.hbs
