+++ packages/sonarwhal/src/lib/cli/wizards/templates/new-parser/script.hbs
