+++ packages/sonarwhal/src/lib/cli/wizards/templates/new-rule/index.ts.hbs
