+++ packages/sonarwhal/src/lib/cli/wizards/templates/new-rule/partial-event-code.hbs
