+++ packages/sonarwhal/src/lib/cli/wizards/templates/new-rule/readme.md.hbs
