+++ packages/sonarwhal/src/lib/cli/wizards/templates/new-rule/tests.ts.hbs
@@ -1,9 +1,8 @@
 import { generateHTMLPage } from 'sonarwhal/dist/tests/helpers/misc';
-import { getRuleName } from 'sonarwhal/dist/src/lib/utils/rule-helpers';
 import { RuleTest } from 'sonarwhal/dist/tests/helpers/rule-test-type';
 import * as ruleRunner from 'sonarwhal/dist/tests/helpers/rule-runner';
 
-const ruleName = getRuleName(__dirname);
+const ruleName = '{{#if parentName}}{{parentName}}/{{/if}}{{normalizedName}}';
 
 /*
  * You should test for cases where the rule passes and doesn't.
