+++ packages/sonarwhal/tests/lib/cli/wizards/fixtures/new.txt
