+++ packages/sonarwhal/tests/lib/cli/wizards/new-rule.ts
@@ -4,7 +4,6 @@ import * as sinon from 'sinon';
 import test from 'ava';
 
 import { CLIOptions } from '../../../../src/lib/types';
-import * as rulesCommon from '../../../../src/lib/cli/rules/common';
 import * as handlebarsUtils from '../../../../src/lib/utils/handlebars';
 
 const actions = ({ newRule: true } as CLIOptions);
@@ -19,16 +18,15 @@ const mkdirp = (dir, callback) => {
     callback();
 };
 
-proxyquire('../../../../src/lib/cli/rules/new-rule', {
+proxyquire('../../../../src/lib/cli/wizards/new-rule', {
     '../../utils/handlebars': handlebarsUtils,
     '../../utils/misc': misc,
-    './common': rulesCommon,
     'fs-extra': fsExtra,
     inquirer,
     mkdirp
 });
 
-import * as rule from '../../../../src/lib/cli/rules/new-rule';
+import * as rule from '../../../../src/lib/cli/wizards/new-rule';
 
 test.serial('If newRule is not an option, it should return false', async (t) => {
     const result = await rule.newRule({} as CLIOptions);
@@ -51,20 +49,18 @@ test.serial('It creates a rule if the option multiple rules is false', async (t)
     const miscWriteFileAsyncStub = sandbox.stub(misc, 'writeFileAsync').resolves();
     const handlebarsCompileTemplateStub = sandbox.stub(handlebarsUtils, 'compileTemplate').returns('');
 
-    sandbox.stub(misc, 'isOfficial').resolves(false);
-    sandbox.stub(rulesCommon, 'processDir').get(() => {
-        return root;
-    });
+    sandbox.stub(misc, 'isOfficial').resolves(true);
+    sandbox.stub(process, 'cwd').returns(root);
     sandbox.stub(inquirer, 'prompt').resolves(results);
 
     const result = await rule.newRule(actions);
 
-    t.true(fsExtraCopyStub.args[0][0].endsWith('files'), 'Unexpected path for external files');
+    t.true(fsExtraCopyStub.args[0][0].endsWith('files'), 'Unexpected path for official files');
     t.is(fsExtraCopyStub.args[0][1], path.join(root, 'rule-awesome-rule'), 'Copy path is not the expected one');
 
-    // changelog, index.ts, license.txt, package.json, readme.md, tsconfig.json, .sonarwhalrc, rule.ts, tests/rule.ts
-    t.is(handlebarsCompileTemplateStub.callCount, 9, `Handlebars doesn't complile the right number of files`);
-    t.is(miscWriteFileAsyncStub.callCount, 9, 'Invalid number of files created');
+    // index.ts, package.json, readme.md, tsconfig.json, rule.ts, tests/rule.ts
+    t.is(handlebarsCompileTemplateStub.callCount, 6, `Handlebars doesn't complile the right number of files`);
+    t.is(miscWriteFileAsyncStub.callCount, 6, 'Invalid number of files created');
 
     t.true(result);
 
@@ -99,9 +95,7 @@ test.serial('It creates a package with multiple rules', async (t) => {
     const handlebarsCompileTemplateStub = sandbox.stub(handlebarsUtils, 'compileTemplate').returns('');
 
     sandbox.stub(misc, 'isOfficial').resolves(false);
-    sandbox.stub(rulesCommon, 'processDir').get(() => {
-        return root;
-    });
+    sandbox.stub(process, 'cwd').returns(root);
     sandbox.stub(inquirer, 'prompt')
         .onFirstCall()
         .resolves(packageResults)
@@ -112,12 +106,14 @@ test.serial('It creates a package with multiple rules', async (t) => {
 
     const result = await rule.newRule(actions);
 
-    t.true(fsExtraCopyStub.args[0][0].endsWith('files'), 'Unexpected path for common files');
+    t.true(fsExtraCopyStub.args[0][0].endsWith('no-official-files'), 'Unexpected path for non official files');
+    t.true(fsExtraCopyStub.args[1][0].endsWith('files'), 'Unexpected path for official files');
     t.is(fsExtraCopyStub.args[0][1], path.join(root, 'rule-awesome-package'), 'Copy path is not the expected one');
+    t.is(fsExtraCopyStub.args[1][1], path.join(root, 'rule-awesome-package'), 'Copy path is not the expected one');
 
-    // changelog, index.ts, license.txt, package.json, readme.md, tsconfig.json, .sonarwhalrc, rule.ts * 2, tests/rule.ts * 2
-    t.is(handlebarsCompileTemplateStub.callCount, 11, `Handlebars doesn't complile the right number of files`);
-    t.is(miscWriteFileAsyncStub.callCount, 11, 'Invalid number of files created');
+    // index.ts, package.json, readme.md, tsconfig.json, .sonarwhalrc, rule.ts * 2, tests/rule.ts * 2
+    t.is(handlebarsCompileTemplateStub.callCount, 9, `Handlebars doesn't complile the right number of files`);
+    t.is(miscWriteFileAsyncStub.callCount, 9, 'Invalid number of files created');
 
     t.true(result);
